move old IllegalMain tests to TestEntryPoint
