+++ main/test/ca/uwaterloo/flix/language/phase/TestTyper.scala
@@ -1145,71 +1145,6 @@ class TestTyper extends FunSuite with TestUtils {
     expectError[TypeError.MismatchedBools](result)
   }
 
-  test("Test.IllegalMain.01") {
-    val input =
-      """
-        |def main(blah: Array[String]): Int32 = ???
-        |""".stripMargin
-    val result = compile(input, Options.TestWithLibMin)
-    expectError[TypeError.IllegalMain](result)
-  }
-
-  test("Test.IllegalMain.02") {
-    val input =
-      """
-        |def main(blah: Array[Char]): Int32 & Impure = ???
-        |""".stripMargin
-    val result = compile(input, Options.TestWithLibMin)
-    expectError[TypeError.IllegalMain](result)
-  }
-
-  test("Test.IllegalMain.03") {
-    val input =
-      """
-        |def main(blah: Array[String]): Int64 & Impure = ???
-        |""".stripMargin
-    val result = compile(input, Options.TestWithLibMin)
-    expectError[TypeError.IllegalMain](result)
-  }
-
-  test("Test.IllegalMain.04") {
-    val input =
-      """
-        |def main(blah: Array[a]): Int32 & Impure = ???
-        |""".stripMargin
-    val result = compile(input, Options.TestWithLibMin)
-    expectError[TypeError.IllegalMain](result)
-  }
-
-  test("Test.IllegalMain.05") {
-    val input =
-      """
-        |class C[a]
-        |
-        |def main(blah: Array[a]): Int32 & Impure with C[a] = ???
-        |""".stripMargin
-    val result = compile(input, Options.TestWithLibMin)
-    expectError[TypeError.IllegalMain](result)
-  }
-
-  test("Test.IllegalMain.06") {
-    val input =
-      """
-        |def main(blah: Array[String]): a & Impure = ???
-        |""".stripMargin
-    val result = compile(input, Options.TestWithLibMin)
-    expectError[TypeError.IllegalMain](result)
-  }
-
-  test("Test.IllegalMain.07") {
-    val input =
-      """
-        |def main(blah: Array[String]): Int32 & ef = ???
-        |""".stripMargin
-    val result = compile(input, Options.TestWithLibMin)
-    expectError[TypeError.IllegalMain](result)
-  }
-
   test("Test.ImpureDeclaredAsPure.01") {
     val input =
       """
